added label and check for unsaved changes on exit
When exiting the application and unsaved changes to the todo file were
made, the user is now asked what should happen with these changes.

Additionally, a "label" is introduced, which indicates the mode the user
is in while typing in the editor.

As a bonus, all "messages" in the editor and labels are extracted and no
longer hardcoded within the used functions.