added highlight for projects, contexts and key-vals
During the implementation of the highlight, I recognized that the
parsing of projects, contexts and key-val pairs (so basically the
'parseAll' function) isn't working properly.

Additionally, the new printing of the item won't retain the original
form of the string. While it would be possible, it would be tedious to
add the highlight in-place and is therefore ditched at the moment.